+++ controller/src/lcd.h
@@ -58,14 +58,14 @@ void lcd_send_string(char *str);
 * 
 * @param data: time in seconds (must be converted )
 */
-void lcd_set_time(int *data);
+void lcd_set_time(uint8_t *data);
 
 /**
 * set temperature to the 3 digits that've been sent over
 * @param mode : using ambient or plant temp
 * @param data : array for temp
 */
-void lcd_set_temperature(uint8_t mode, int *data);
+void lcd_set_temperature(uint8_t mode, uint8_t *data);
 
 /**
 * toggle cursor on lcd
